FAPI: Fix deviation from CEL specification.
"template_value" was used as field name for IMA json events.
The notion template value was used in older versions of the
CEL spec. In the current version (v1_r0p38) the name
"template_data" is used.
To achieve compatibility with TSS version 4.0. "template_value"
is still allowed in IMA logs. In the output "template_data"
will be used.

Signed-off-by: Juergen Repp <juergen_repp@web.de>