FAPI: Fix missing parameter encryption for policy sessions.
The parameter encryption for policy sessions was not enabled.
Now the parameter encryption is enabled and the auth value of
objects is added to the session key.
One exception is the cp hash policy. In this case the
the cp hash check forced by this policy would fail with an
encrypted parameter.

Signed-off-by: Juergen Repp <juergen_repp@web.de>